fix(bl31): remove incorrect asserts
When RESET_TO_BL31 is defined, BL31 is the reset vector and no
parameters are provided by a previous bootloader.
Therefore, the arguments arg0 to arg3 passed to
arm_bl31_early_platform_setup() must be ignored.
They come directly from the BL31 entrypoint.
The GP registers reset to an architecturally unknown value
so are not necessarily initialized to zero.

Signed-off-by: Alexandre Gonzalo <alexandre.gonzalo@arm.com>
Change-Id: I38dde35145f2bd27fe9645906df192b138984c1d