fix(versal): runtime console in debug mode
Whenever DEBUG is set to 1, the Makefile forces CONSOLE_RUNTIME to
pl011, regardless of the user-specified VERSAL_CONSOLE value. This
causes a build requested with VERSAL_CONSOLE=pl011_1 to register both
pl011_1 and pl011 as boot and runtime consoles. If the hardware is
connected only to UART1, this causes TF-A to hang when UART0 is
selected as the runtime console, since it waits indefinitely on the
transmit FIFO. The fix ensures that, in a DEBUG build, CONSOLE_RUNTIME
defaults to the same value as VERSAL_CONSOLE.

Change-Id: I7aeedb04040ea4ab4b8aecde98af5cc39df09c1a
Signed-off-by: Prasad Kummari <prasad.kummari@amd.com>