fix(bl2): unify the BL2 EL3 and RME entrypoints
BL2 has 3(!) entrypoints:
 1) the regular EL1 entrypoint (once per AArch)
 2) an EL3 entrypoint
 3) an EL3 entrypoint with RME

The EL1 and EL3 entrypoints are quite distinct so it's useful to keep
them separate. But the EL3 and RME entrypoints are conceptually
identical just configured differently and having slightly different
assumptions (eg whether we can rely on BL1). So put them together with
only the configuration as a difference. This has a few benefits:
 * makes the naming consistent - BL2 always runs at EL1, BL2_EL3 always
   runs at EL3. This is most important for the linker script.
 * paves the way for ENABLE_RME and RESET_TO_BL2 to coexist.
 * allows for more general refactors

Currently, ENABLE_RME and RESET_TO_BL2 are mutually exclusive (from a
makefile constraint) so the checks are simplified to one or the other as
there is no danger of their simultaneous use.

Change-Id: Iecffab2ff3a0bd7823f8277d9f66e22e4f42cc8c
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>