fix(trp): report argument errors before bailing out
TRP checks for the extra warmboot arguments (in registers x2 and x3) to
be 0, then bails out of the RME initialisation process if they are not.
However it does not report about that case, so the users will be left
clueless about what is going wrong.

Add an error message reporting the fact and the values of the registers,
when they are not 0.

Change-Id: I6231551ccdfa9434d0153caa95933044e2b23938
Signed-off-by: Andre Przywara <andre.przywara@arm.com>