fix(cpus): use correct Makefile indentation for CVE-2018-3639 check
Makefiles need to use spaces for indentation when using make syntax,
tabs are reserved for (shell) recipes.

Replace tabs with spaces on the WORKAROUND_CVE_2018_3639 check, to fix
the error report when WORKAROUND_CVE_2018_3639 is disabled:
lib/cpus/cpu-ops.mk:1147: *** recipe commences before first target. Stop.

Also this revealed that DYNAMIC_WORKAROUND_CVE_2018_3639 was not
initialised, so it always triggered that condition. Set it to 0, to
allow disabling WORKAROUND_CVE_2018_3639 on the command line.

Use the opportunity to also convert some unrelated tab to spaces, in a
line continuation.

Change-Id: Ieb56af33a11c40b6753738669eee929c264261cf
Signed-off-by: Andre Przywara <andre.przywara@arm.com>