fix(build): prevent races on the build directory
The tools (eg fiptool) don't depend on the build directory so it's
possible that make tries to build them before it exists. Doing that
leads to random and unpredictable errors. Almost always, they are built
after a BL image which always has a build directory dependency, but
when building MANY tf-a builds over MANY threads concurrently this could
be observed (I suspect the high load of the system increases latency
just enough that this race is lost).

The fix is simple - have an explicit dependency on the build directory.

This is the same problem as 9855568cc and 25cde5f81.

Change-Id: I769ac07f8882f82ea9d72f3b976337284d697310
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>