feat(cpus): add pabandon support to Nevis
Nevis' TRM says that a powerdown attempt may be abandoned for a handful
of reasons. Add support for handling this.

It also says that if the SME2 engine is not properly disconnected, then
a powerdown request will be rejected. Require ERRATA_SME_POWER_DOWN be
set to avoid this.

Just like Gelas/Travis, the 11.28 model doesn't reset the bit, so a
workaround is necessary.

Change-Id: I0d5de1b0e772f6b4d656e841fb4bcf8fd859f293
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>