fix(build): correctly forward C compiler to CMake
When `$($(ARCH)-cc)` does not represent a valid program name or path,
CMake fails to configure the Event Log library.

This occurs because CMake treats `CMAKE_C_COMPILER` as a
semicolon-delimited list of arguments to pass to `exec()`, whereas we
provide it the value of `$($(ARCH)-cc)`, which is a shell fragment that
does not necessarily represent a correct program path.

This change ensures that we pass a semicolon-delimited list as is
expected by CMake.

Change-Id: I1dd3e53624054c4b5c881be1ab99e77ea22ceba2
Signed-off-by: Chris Kay <chris.kay@arm.com>