fix(build): force CMake to cross-compile
CMake provides the `CMAKE_SYSTEM_NAME` variable, which identifies the
target operating system that the project is being built for. By default
this variable inherits its value from `CMAKE_HOST_SYSTEM_NAME`, which
identifies the operating system of the host machine.

If these two variables differ in value, CMake determines that the
project is being cross-compiled, and changes some of its behaviours.

For example: on macOS, unless cross-compiling is enabled, CMake passes
Apple Clang's `-arch` flag to the C compiler whether it is Apple Clang
or not.

To resolve this, this change now sets the following CMake variables:

  - `CMAKE_SYSTEM_NAME` to `Generic`, indicating no specific system.
  - `CMAKE_SYSTEM_VERSION` to ``, to un-inherit the host system version.

See the CMake documentation for the following variables:

  - https://cmake.org/cmake/help/latest/variable/CMAKE_LANG_COMPILER.html
  - https://cmake.org/cmake/help/latest/variable/CMAKE_LANG_HOST_COMPILER.html
  - https://cmake.org/cmake/help/latest/variable/CMAKE_CROSSCOMPILING.html

Change-Id: Id5406c61baec459f56b6a105b3b507d238ff2847
Signed-off-by: Chris Kay <chris.kay@arm.com>