feat(build): add helpers for managing compiler launchers
With the introduction of the Event Log library, we now need to forward
some of our toolchain information to CMake. However, CMake always treats
the first shell word in `CC` as the path to the compiler, which is
problematic for users of compiler launchers like `ccache`.

This change introduces some additional heuristics to the toolchain
interface to detect compiler launchers and isolate their arguments from
the underlying compiler command, which can then be forwarded separately
to CMake.

Change-Id: I783833c523380b989bbac05b57503bc76881f6a4
Signed-off-by: Chris Kay <chris.kay@arm.com>