refactor(build): unify toolchain derivation
At the moment, the process we go through to derive the value for the
toolchain tool variables is fairly complicated and not particularly
consistent. This change brings the three mechanisms we support into
alignment in terms of their inputs and outputs via three helpers:

- `toolchain-from-parameter`: load the command from its parameter
- `toolchain-from-cc`: load the command from its C compiler
- `toolchain-from-default`: load the command from its built-in defaults

These functions take a toolchain identifier and a tool class identifer,
and return either a program name/path or a shell command. Sanitisation
is then done inside `toolchain-determine-tool`, which now makes a
stronger guarantee that the toolchain tools are correctly escaped for
the shell.

Change-Id: I3ce73134824e805e8ee47a119c982e7fe4923fae
Signed-off-by: Chris Kay <chris.kay@arm.com>