chore(xilinx): add deprecation warning to pm_feature_check
Add deprecation warning to pm_feature_check() function to inform
users that this API will be removed in the 2027.1 release and
they should migrate to tfa_api_feature_check() for TF-A specific
feature checks.

This warning helps customers prepare for the upcoming API removal
and encourages migration to the correct function.

Change-Id: Icab5eb6f1a552553b1cc1215aa683430733667bd
Signed-off-by: Devanshi Chauhan <devanshi.chauhan@amd.com>