docs: add Architectural Feature Support reference
Introduce a central document to track the status of Arm architectural
features in TF-A.

This aims to provide a single reference point to check whether a given
feature is:
 - explicitly supported in TF-A (OK)
 - transparent from EL3 (no changes required) (NA)
 - Analyzed but decided not to implemened (NO)
 - Implementation in progress (WIP)
 - not yet analyzed.

This reduces the current reliance on grepping the code, browsing JIRA,
or cross-referencing the Arm ARM to answer feature status queries.

The content is aligned with Arm’s yearly architectural feature updates
(see [Architecture Feature Descriptions](https://developer.arm.com/documentation/109697/latest/)).

Signed-off-by: Manish Pandey <manish.pandey2@arm.com>
Change-Id: I32cc07268fc641180837a42a973308dab0824236