feat(el3-runtime): implement synchronous EA injection
External aborts (EAs) can be synchronous or asynchronous and both can be
routed to EL3 via SCR_EL3.EA (via FFH_SUPPORT). When this happens, the
platform should perform its handling and return the error back to where
it was caused. As this is a synchronous exception, it is not maskable
and its target is easily computed so injection is straightforward by
following the exception taking code from the architecture.

This allows for SDEI to not be used for synchronous EAs. The problem
with SDEI is that it must be implemented separately on each platform and
is a very heavy protocol, impacting code size, runtime performance, and
general maintenance burden.

Change-Id: Ia55fbb92d68fd1568020845759649cfd791cf801
Co-developed-by: Venkata Pokkuluri <vpokkuluri@google.com>
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>